+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - docs: fix section level in README
 - refactor: extract command classes
+- feat: load token from configuration file
 
 ## [v2.2.0] - 2025-01-10